intel/openvino_2019.3.376/deployment_tools/tools/workbench/requirements.txt
celery==4.3
defusedxml>=0.6.0
eventlet>=0.24.1
Flask>=1.0.2
Flask-RESTful>=0.3.6
Flask-SocketIO>=3.0.2
Flask-SQLAlchemy>=2.3.2
gunicorn>=19.9.0
greenlet>=0.4.15
numpy>=1.14.5
psycopg2-binary==2.8.2
pyyaml>=5.1